changed view modes providers localisation
now in appropriate view models
