+++ pwi_project/lib/view/notepad_screen_view.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:pwi_project/utils/text_field_controllers.dart';
-import 'package:pwi_project/utils/view_modes.dart';
 import 'package:pwi_project/view_model/note_view_model.dart';
 import 'package:pwi_project/view_model/notepad_view_model.dart';
