+++ pwi_project/lib/widgets/notestick_grid.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:pwi_project/utils/view_modes.dart';
 import 'package:pwi_project/view/notepad_screen_view.dart';
+import 'package:pwi_project/view_model/notepad_view_model.dart';
 
 import '../view_model/note_view_model.dart';
 import 'note_stick.dart';
